requirements.txt
plotly==3.7.0
python-dotenv==0.10.1
requests==2.21.0
